backend/python_project/pyproject.toml
[tool.poetry]
name = "python-project"
version = "0.1.0"
description = ""
authors = ["Antonio Dias <antonio_g_d@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-cors = "^5.0.0"
python-dotenv = "^1.0.1"
flask-pydantic-spec = "^0.6.0"
sqlalchemy = "^2.0.36"
flask-sqlalchemy = "^3.1.1"
flask-testing = "^0.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
